# OR_Labos/openapi.json
{
	"openapi": "3.0.3",
	"info": {
		"description": "OpenAPI specifikacija za bazu podataka o pivama.",
		"version": "1.0.0",
		"title": "Pive API",
		"contact": {
			"email": "is52364@fer.hr"
		},
		"license": {
			"name": "Creative Commons Zero v1.0 Universal",
			"url": "https://creativecommons.org/publicdomain/zero/1.0/"
		}
	},
	"host": "localhost:8080",
	"paths": {
		"/datatable/data": {
			"get": {
				"tags": [
					"datatable"
				],
				"summary": "Dohvati sve podatke iz baze",
				"description": "Dohvatit će se svi podaci o pivama koji su zapisani u bazi podataka",
				"operationId": "getData",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Dohvaćeni podaci o pivi",
						"schema": {
							"$ref": "#/definitions/PiveResponse"
						}
					},
					"404": {
						"description": "Podaci nisu pronađeni!"
					}
				}
			}
		},
		"/datatable/details/{idpive}": {
			"get": {
				"tags": [
					"datatable"
				],
				"summary": "Dohvati sve podatke iz baze",
				"description": "Dohvatit će se svi podaci o piavama koji su zapisani u bazi podataka",
				"operationId": "showDetails",
				"parameters": [
					{
						"name": "idpive",
						"in": "path",
						"description": "id pive čiji detalji se traže",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Dohvaćeni podaci o pivi",
						"schema": {
							"$ref": "#/definitions/PivaResponse"
						}
					},
					"404": {
						"description": "Podaci nisu pronađeni!"
					}
				}
			}
		},
		"/datatable/new": {
			"post": {
				"tags": [
					"datatable"
				],
				"summary": "Dodaj pivu",
				"description": "Dodaj novi zapis u bazu",
				"operationId": "addNewEntry",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "Objekt s podacima o pivi",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Piva"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Dodan novi zapis",
						"schema": {
							"$ref": "#/definitions/IdResponse"
						}
					}
				}
			}
		},
		"/datatable/delete/{idpive}": {
			"delete": {
				"tags": [
					"datatable"
				],
				"summary": "Briše pivu",
				"description": "Briše zapis iz baze podataka",
				"operationId": "deleteEntry",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "idpive",
						"in": "path",
						"description": "id pive koju treba izbrisati",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "Podatak je izbrisan!",
						"schema": {
							"$ref": "#/definitions/IdResponse"
						}
					},
					"404": {
						"description": "Podatak nije pronađen!"
					}
				}
			}
		},
		"/datatable/edit/{idpive}": {
			"put": {
				"tags": [
					"datatable"
				],
				"summary": "Uređuje pivu",
				"description": "Uređuje zapis iz baze podataka",
				"operationId": "editEntry",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "idpive",
						"in": "path",
						"description": "id pive koju treba urediti",
						"required": true,
						"type": "integer",
						"format": "int64"
					},
					{
						"in": "body",
						"name": "body",
						"description": "Objekt s podacima o pivi",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Piva"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Podatak je uređen!",
						"schema": {
							"$ref": "#/definitions/IdResponse"
						}
					},
					"404": {
						"description": "Podatak nije pronađen!"
					}
				}
			}
		},
		"/datatable/pive/nazivi": {
			"get": {
				"tags": [
					"datatable"
				],
				"summary": "Dohvati nazive pivi iz baze",
				"description": "Dohvatit će se svi nazivi o piva koji su zapisani u bazi podataka",
				"operationId": "showTitles",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Podaci su pronađeni!i",
						"schema": {
							"$ref": "#/definitions/NaziviResponse"
						}
					},
					"404": {
						"description": "Podaci nisu pronađeni!"
					}
				}
			}
		},
		"/datatable/pive/pakiranja": {
			"get": {
				"tags": [
					"datatable"
				],
				"summary": "Dohvati pakiranja iz paze",
				"description": "Dohvatit će se sva pakiranja u kojima može biti piva koja su zapisana u bazi podataka",
				"operationId": "showPackaging",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Podaci su pronađeni!",
						"schema": {
							"$ref": "#/definitions/PakiranjaResponse"
						}
					},
					"404": {
						"description": "Podaci nisu pronađeni!"
					}
				}
			}
		},
		"/datatable/pive/last": {
			"get": {
				"tags": [
					"datatable"
				],
				"summary": "Dohvati zadnju pivu iz paze",
				"description": "Dohvatit će se zadnji zapis iz baze podataka",
				"operationId": "showLastEntry",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Podaci su pronađeni!",
						"schema": {
							"$ref": "#/definitions/PivaResponse"
						}
					},
					"404": {
						"description": "Podaci nisu pronađeni!"
					}
				}
			}
		}
	},
	"definitions": {
		"Piva": {
			"type": "object",
			"required": [
				"idpive",
				"postalkohola",
				"boja",
				"vrsta",
				"distributerzarh",
				"drzporjekla",
				"envrijednost100ml",
				"idpakiranja",
				"materijal",
				"velicina"
			],
			"properties": {
				"id": {
					"type": "integer",
					"format": "int64",
					"description": "Id pive"
				},
				"postalkohola": {
					"type": "string",
					"description": "Postotak alkohola"
				},
				"boja": {
					"type": "string",
					"description": "Boja pive"
				},
				"vrsta": {
					"type": "string",
					"description": "Vrsta pive"
				},
				"distributerzarh": {
					"type": "string",
					"description": "Distributer pive za Hrvatsku"
				},
				"drzporjekla": {
					"type": "string",
					"description": "Država porjekla pive"
				},
				"envrijednost100ml": {
					"type": "string",
					"description": "Energetska vrijednost pive sadržana u 100 ml"
				},
				"idpakiranja": {
					"type": "integer",
					"format": "int64",
					"description": "Id vrste pakiranja u kojoj je piva"
				},
				"materijal": {
					"type": "string",
					"description": "Materijal vrste pakiranja"
				},
				"velicina": {
					"type": "string",
					"description": "Veličina vrste pakiranja"
				}
			}
		},
		"PiveResponse": {
			"type": "object",
			"properties": {
				"status": {
					"type": "integer",
					"format": "int64"
				},
				"message": {
					"type": "string"
				},
				"response": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Piva"
					}
				}
			}
		},
		"PivaResponse": {
			"type": "object",
			"properties": {
				"status": {
					"type": "integer",
					"format": "int64"
				},
				"message": {
					"type": "string"
				},
				"response": {
					"$ref": "#/definitions/Piva"
				}
			}
		},
		"IdResponse": {
			"type": "object",
			"properties": {
				"status": {
					"type": "integer",
					"format": "int64"
				},
				"message": {
					"type": "string"
				},
				"response": {
					"type": "integer",
					"format": "int64"
				}
			}
		},
		"NaziviResponse": {
			"type": "object",
			"properties": {
				"status": {
					"type": "integer",
					"format": "int64"
				},
				"message": {
					"type": "string"
				},
				"response": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			}
		},
		"Pakiranje": {
			"type": "object",
			"required": [
				"idpakiranja",
				"materijal",
				"velicina"
			],
			"properties": {
				"idpakiranja": {
					"type": "integer",
					"format": "int64",
					"description": "Id vrste pakiranja u kojoj je piva"
				},
				"materijal": {
					"type": "string",
					"description": "Materijal vrste pakiranja"
				},
				"velicina": {
					"type": "string",
					"description": "Veličina vrste pakiranja"
				}
			}
		},
		"PakiranjaResponse": {
			"type": "object",
			"properties": {
				"status": {
					"type": "integer",
					"format": "int64"
				},
				"message": {
					"type": "string"
				},
				"response": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Pakiranje"
					}
				}
			}
		}
	}
}